Move database to entity folder
